# Speeds up fallback to Hash#default_proc in rb_hash_aref by removing a method call